+++ src/css/custom.css
@@ -147,7 +147,7 @@ a{
   padding: 25px;
   background-color: transparent;
   border-radius: 0;
-  border: solid 2px #19af54;
+  border: solid 1px #000000;
   color: #19af54;
   cursor: pointer;
   transition: all .5s ease;
@@ -965,6 +965,15 @@ a{
   margin-bottom: 3rem;
 }
 
+.boxIconButton{
+  padding-left: 0;
+  padding-right: 0;
+  padding-top: 13px;
+  padding-bottom: 13px;
+  margin-top: 45px;
+  width: 200px
+}
+
 .buttonBoxMobile{
   margin-top: 50px;
   width: 100%;
@@ -1585,6 +1594,16 @@ a{
   }
 
   .buttonBox{
+    width: 100%;
+    display: flex;
+    justify-content: center;
+  }
+
+  .boxIconButton{
+    margin-top: 0;
+  }
+
+  .specialButtonBox{
     width: 100%;
     display: flex;
     justify-content: center;
